requirements.txt
yfinance
pandas
numpy
alpaca-py
python-dotenv
